Initial commit with base library and markov-chain binary.
The markov chain library is relatively short right now, and only has a
one special impl for strings on the chain data structure. Serialization
is supported by default, and may be turned off in compilation.

There's also a management program that will create markov chains in the
spcecified format, and read them as well. This is useful for inspecting
chains and looking at what they have inside of them. The management
program is not complete yet, but it is able to save markov chains to
files and update them.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>