Add generate subcommand to markov-chain utility, and merge.
Generate is the big change to the markov-chain utility. It allows many
files that may or may not be markov chains, and will merge a chain
together on the fly to generate some sentences.

Merge is the big internal change, which allows for merging of markov
chains. This will lead into the merge utility subcommand.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>