Update the built-without-generator message.
The markov-util generator is an optional feature, and if the program is built
without the "generator" feature, the binary will print a message telling the
user which flags to use to build the generator as well. This used to be
determined by the "clap" package, but now has its own custom feature name;
this has been updated to reflect the changes.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>